release v4.2 to devel

Closes #1

See merge request uib/python-opsi!995

+++ run_tests.sh
@@ -1,6 +1,6 @@
 #! /bin/bash
-if [ -e coverage.xml ]; then
-	rm coverage.xml
-fi
+[ -e coverage.xml ] && rm coverage.xml
+set -e
 
-py.test --junitxml=testreport.xml --cov-config .coveragerc --cov OPSI --cov-report xml --quiet tests/
+py.test --log-level=WARNING -o junit_family=xunit2 --junitxml=testreport-OPSI.xml --cov-append --cov OPSI --cov-report term --cov-report xml -v tests/testOPSI
+py.test --log-level=WARNING -o junit_family=xunit2 --junitxml=testreport-opsicommon.xml --cov-append --cov opsicommon --cov-report term --cov-report xml -v tests/testopsicommon
